+++ hot-loader-helper.js
@@ -1,9 +1,9 @@
-module.exports.reload = function(type) { return `
+module.exports.reload = function ({ type, module }) {
+    return `
     if (module.hot) {
         module.hot.accept();
         module.hot.dispose(() => {
-            global.__hmrRefresh('${type}');
+            global.__hmrRefresh({ type: '${type}', module: '${module}' });
         })
     }
 `};
-