+++ markup-hot-loader.js
@@ -1,5 +1,7 @@
 const { reload } = require("./hot-loader-helper");
 
 module.exports = function (source) {
-    return `${source};${reload('markup')}`;
+    const typeMarkup = "markup";
+    const modulePath = this.resourcePath.replace(this.context, ".");
+    return `${source};${reload({ type: typeMarkup, module: modulePath })}`;
 };